==> 0-Titanics-Preparring for Models.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

df_train = pd.read_csv('datasets/train.csv')
df_test = pd.read_csv('datasets/test.csv')

## 1. Overview

df_train.head()

df_train.info()

df_test.head()

df_test.info()

df_train["PassengerId"].duplicated().sum(), df_test["PassengerId"].duplicated().sum()

df_train[["Age", "Fare"]].describe()

pd.concat([df_train.Pclass.astype('category'), 
           df_train.SibSp.astype('category'), 
           df_train.Parch.astype('category'),
           df_train.select_dtypes(include ='O')
          ], axis = 1).describe()

## 2. Xoá cột không cần thiết

def drop_colums(df, cols):
    for el_df in df:
        el_df.drop(cols, axis = 1, inplace = True)
# cols = ['Name', 'Ticket']
save_PassengerId_test = df_test.PassengerId 
drop_colums(df = [df_train, df_test], cols = ['PassengerId', 'Name', 'Ticket', 'Cabin', 'Embarked'])

#df_train.info(),  df_test.info(), save_PassengerId_test.head()

df_train.info()

## 3. Preparing Data for modelling

### a. Columns Object_category to dummies (không cần cho vào pipeline)

df_train = pd.get_dummies(df_train, drop_first = True)
df_test = pd.get_dummies(df_test, drop_first = True)

df_train.info()

### b. To save

df_train.to_csv('datasets_for_preprocessing/train_for_pipeline.csv', index=False)
df_test.to_csv('datasets_for_preprocessing/test_for_pipeline.csv', index=False)
