==> 1-Linear models.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')
%cd /content/gdrive/MyDrive/Andrew-PROJECT/titanic

# 1. Load Data

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import pickle

train_data = pd.read_csv('train_for_pipeline.csv')
X_test = pd.read_csv('test_for_pipeline.csv')

X_train = train_data.drop("Survived", axis = 1)
y_train = train_data["Survived"]

df_test = pd.read_csv('test.csv')

# 2. Logistic Regression model

from sklearn.impute import SimpleImputer
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.model_selection  import StratifiedKFold
from sklearn.preprocessing import StandardScaler, MinMaxScaler

cv = StratifiedKFold(n_splits=3, shuffle=False, random_state=0)

logreg = LogisticRegression()
imp = SimpleImputer(strategy = 'mean')
#scaler = StandardScaler()
scaler = MinMaxScaler()
steps = [('imp', imp), ("scaler", scaler),('logreg', logreg)]
pipe = Pipeline(steps)

param_grid = [
    {'logreg__penalty': ['l1'], 'logreg__solver': ['liblinear'],
         'logreg__C':[0.1, 0.2, 0.5, 1, 1.1, 1.2, 1.3, 1.4, 1.5, 2, 2.5, 50], 'logreg__max_iter':[100, 200] },
    {'logreg__C':[0.1, 0.2, 0.5, 1, 1.1, 1.2, 1.3, 1.4, 1.5, 2, 2.5, 50], 'logreg__max_iter':[100, 200]}
]
logreg_gridcv = GridSearchCV(pipe, param_grid = param_grid, cv=cv, scoring="accuracy")
logreg_gridcv.fit(X_train, y_train) 

logreg_gridcv.best_score_, logreg_gridcv.best_params_

logreg_gridcv.score(X_train,y_train)

# Enregistrement
pickle.dump(logreg_gridcv, open('logreg_cv.sav', 'wb'))

results = pd.DataFrame(logreg_gridcv.cv_results_)
print(results[['params',  'mean_test_score']])


y_pred = logreg_gridcv.predict(X_test)
df_result = pd.DataFrame({
        "PassengerId": df_test["PassengerId"],
        "Survived": y_pred})
df_result.to_csv('reluts_logreg.csv', index = False)

# 3. Linear SVC

from sklearn.svm import LinearSVC
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.impute import SimpleImputer



imp_svc = SimpleImputer(strategy = 'mean')
#scaler_svc = StandardScaler()
scaler_svc = MinMaxScaler()
svc = LinearSVC()

steps_svc = [('imp_svc', imp_svc), ('scaler', scaler_svc), ('svc', svc)]
pipe_svc = Pipeline(steps_svc)

param_grid_svc = {'svc__C': [0.1, 0.15, 0.2, 0.3, 0.5, 1, 2, 3, 4, 5, 6, 7, 8, 9, 10]}
svc_gridcv = GridSearchCV(pipe_svc, param_grid = param_grid_svc, cv=cv, scoring="accuracy")

svc_gridcv.fit(X_train, y_train)



svc_gridcv.best_params_, svc_gridcv.score(X_train, y_train)

y_pred = svc_gridcv.predict(X_test)
df_result = pd.DataFrame({
        "PassengerId": df_test["PassengerId"],
        "Survived": y_pred})
df_result.to_csv('reslut_svc.csv', index = False)

# 4. KNN model

from sklearn.neighbors import KNeighborsClassifier
from sklearn.preprocessing import StandardScaler
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV
from sklearn.impute import SimpleImputer

imp_knn = SimpleImputer(strategy = 'mean')
scaler_knn = StandardScaler()
#scaler_knn = MinMaxScaler()
knn = KNeighborsClassifier()

steps_knn = [('imp', imp_knn), ('scaler', scaler_knn), ('knn', knn)]
pipe_knn = Pipeline(steps_knn)
param_grid_knn = {'knn__n_neighbors': [1, 2, 3, 4, 5, 6, 7, 8, 9, 10]}
knn_gridcv = GridSearchCV(pipe_knn, param_grid = param_grid_knn, cv=cv, scoring="accuracy")

knn_gridcv.fit(X_train, y_train)


knn_gridcv.best_params_
knn_gridcv.best_score_

knn_gridcv.score(X_train, y_train)

y_pred = knn_gridcv.predict(X_test)

# 100. OUTPUT


df_test = pd.read_csv('test.csv')

df_test.head()

df_result = pd.DataFrame({
        "PassengerId": df_test["PassengerId"],
        "Survived": y_pred})
df_result.to_csv('reluts_knn.csv', index = False)
