==> 2-Knn and SVM.ipynb <==
from google.colab import drive
drive.mount('/content/gdrive')
%cd /content/gdrive/MyDrive/AI/Projects/Titanic

# 1. Load Data

import numpy as np
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns

import pickle

train_data = pd.read_csv('datasets_for_preprocessing/train_for_pipeline.csv')
X_test = pd.read_csv('datasets_for_preprocessing/test_for_pipeline.csv')

X_train = train_data.drop("Survived", axis = 1)
y_train = train_data["Survived"]

df_test = pd.read_csv('datasets/test.csv')

# II. Our two models using Grid Search

from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, MinMaxScaler
from sklearn.decomposition import PCA

from sklearn.svm import SVC
from sklearn.neighbors import KNeighborsClassifier

from sklearn.pipeline import Pipeline

from sklearn.model_selection import GridSearchCV
from sklearn.model_selection  import StratifiedKFold

cv = StratifiedKFold(n_splits=3, shuffle=False)

# 1. Knn

imp = SimpleImputer(strategy = 'mean')
scaler = StandardScaler()
#scaler = MinMaxScaler()
#pca = PCA()
model = KNeighborsClassifier()

steps = [('imp', imp), ("scaler", scaler),('model', model)]
pipe = Pipeline(steps)

param_grid = {"model__n_neighbors":[2,3,4,5,6,7,8,9,10]}


model_gridcv = GridSearchCV(pipe, param_grid = param_grid, cv=cv, scoring="accuracy")

model_gridcv.fit(X_train, y_train)

model_gridcv.best_params_, 

model_gridcv.best_score_, model_gridcv.score(X_train, y_train)

pd.DataFrame(model_gridcv.cv_results_)[['params','mean_test_score' ]].sort_values("mean_test_score", ascending = False)

def for_kaggle_and_save_model(model, model_gridcv):
  # a. For Kaggle
  y_pred = model_gridcv.predict(X_test)
  df_result = pd.DataFrame({
        "PassengerId": df_test["PassengerId"],
        "Survived": y_pred})
  df_result.to_csv("datasets_results/" + model.__class__.__name__ +".csv", index = False)
  # b. Save the model
  pkl_filename = "saved_models/" + model.__class__.__name__  + ".pkl"
  with open(pkl_filename, 'wb') as file:
    pickle.dump(model_gridcv, file)

for_kaggle_and_save_model(model, model_gridcv)

## 2. Kernel SVM

#imp = SimpleImputer(strategy = 'mean')
#scaler = StandardScaler()
scaler = MinMaxScaler()
svc = SVC()

steps = [('imp', imp), ('scaler', scaler), ('svc', svc)]
pipe = Pipeline(steps)



param_grid = {'svc__kernel': ['rbf', "sigmoid"], 'svc__C': [0.1, 0.15, 0.2, 0.3, 0.5, 1, 2, 2.5, 3, 3.5, 4, 5, 6, 7, 8, 9, 10]}
svc_gridcv = GridSearchCV(pipe, param_grid = param_grid, cv=cv, scoring="accuracy")

svc_gridcv.fit(X_train, y_train)

svc_gridcv.best_params_, 

svc_gridcv.best_score_, svc_gridcv.score(X_train, y_train)

pd.DataFrame(svc_gridcv.cv_results_)[['params','mean_test_score' ]].sort_values("mean_test_score", ascending = False).head()

for_kaggle_and_save_model(svc, svc_gridcv)
